restructured folder structure and imports
